# openapi_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "\u623f\u5c4b\u79df\u8d41AI\u8425\u9500\u7cfb\u7edf",
        "description": "AI\u9a71\u52a8\u7684\u623f\u6e90\u6587\u6848\u751f\u6210\u4e0e\u591a\u5e73\u53f0\u53d1\u5e03\u7cfb\u7edf",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/houses/upload": {
            "post": {
                "tags": [
                    "\u623f\u6e90\u7ba1\u7406"
                ],
                "summary": "Upload House",
                "description": "\u4e0a\u4f20\u623f\u6e90\u56fe\u7247\u548c\u57fa\u672c\u4fe1\u606f\n\n- **images**: \u623f\u6e90\u56fe\u7247\u6587\u4ef6\u5217\u8868\uff08multipart/form-data\uff09\n- **house_info**: \u623f\u6e90\u57fa\u672c\u4fe1\u606f\uff08JSON\u5b57\u7b26\u4e32\uff0c\u53ef\u9009\uff09\n\n\u8fd4\u56de\uff1a\u521b\u5efa\u7684\u623f\u6e90\u5bf9\u8c61",
                "operationId": "upload_house_api_v1_houses_upload_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_house_api_v1_houses_upload_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HouseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/houses": {
            "get": {
                "tags": [
                    "\u623f\u6e90\u7ba1\u7406"
                ],
                "summary": "Get Houses",
                "description": "\u83b7\u53d6\u623f\u6e90\u5217\u8868\n\n- **skip**: \u8df3\u8fc7\u8bb0\u5f55\u6570\uff08\u5206\u9875\uff09\n- **limit**: \u8fd4\u56de\u8bb0\u5f55\u6570\uff08\u5206\u9875\uff09\n\n\u8fd4\u56de\uff1a\u623f\u6e90\u5217\u8868\u548c\u603b\u6570",
                "operationId": "get_houses_api_v1_houses_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HouseListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/houses/{house_id}": {
            "get": {
                "tags": [
                    "\u623f\u6e90\u7ba1\u7406"
                ],
                "summary": "Get House",
                "description": "\u83b7\u53d6\u623f\u6e90\u8be6\u60c5\n\n- **house_id**: \u623f\u6e90ID\n\n\u8fd4\u56de\uff1a\u623f\u6e90\u8be6\u7ec6\u4fe1\u606f",
                "operationId": "get_house_api_v1_houses__house_id__get",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "House Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HouseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "\u623f\u6e90\u7ba1\u7406"
                ],
                "summary": "Delete House",
                "description": "\u5220\u9664\u623f\u6e90\n\n- **house_id**: \u623f\u6e90ID\n\n\u8fd4\u56de\uff1a204 No Content",
                "operationId": "delete_house_api_v1_houses__house_id__delete",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "House Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/history": {
            "get": {
                "tags": [
                    "\u5386\u53f2\u8bb0\u5f55"
                ],
                "summary": "Get History",
                "description": "\u83b7\u53d6\u5386\u53f2\u8bb0\u5f55\uff08\u6574\u5408\u623f\u6e90\u3001\u6587\u6848\u3001\u53d1\u5e03\u8bb0\u5f55\uff09\n\n- **skip**: \u8df3\u8fc7\u8bb0\u5f55\u6570\n- **limit**: \u8fd4\u56de\u8bb0\u5f55\u6570\n\n\u8fd4\u56de\uff1a\u5386\u53f2\u8bb0\u5f55\u5217\u8868\uff08\u5305\u542b\u623f\u6e90\u3001\u6587\u6848\u3001\u53d1\u5e03\u72b6\u6001\uff09",
                "operationId": "get_history_api_v1_history_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    },
                                    "title": "Response Get History Api V1 History Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/history/{house_id}": {
            "delete": {
                "tags": [
                    "\u5386\u53f2\u8bb0\u5f55"
                ],
                "summary": "Delete History",
                "description": "\u5220\u9664\u5386\u53f2\u8bb0\u5f55\uff08\u540c\u65f6\u5220\u9664\u623f\u6e90\u3001\u6587\u6848\u3001\u53d1\u5e03\u8bb0\u5f55\uff09\n\n- **house_id**: \u623f\u6e90ID\n\n\u8fd4\u56de\uff1a204 No Content",
                "operationId": "delete_history_api_v1_history__house_id__delete",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "House Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_upload_house_api_v1_houses_upload_post": {
                "properties": {
                    "images": {
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "type": "array",
                        "title": "Images",
                        "description": "\u623f\u6e90\u56fe\u7247\uff08\u591a\u5f20\uff09"
                    },
                    "house_info": {
                        "type": "string",
                        "title": "House Info",
                        "description": "\u623f\u6e90\u4fe1\u606f\uff08JSON\u683c\u5f0f\uff09"
                    }
                },
                "type": "object",
                "required": [
                    "images"
                ],
                "title": "Body_upload_house_api_v1_houses_upload_post"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HouseListResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/HouseResponse"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total"
                ],
                "title": "HouseListResponse",
                "description": "\u623f\u6e90\u5217\u8868\u54cd\u5e94Schema"
            },
            "HouseResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    },
                    "rent": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rent"
                    },
                    "rooms": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rooms"
                    },
                    "area": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Area"
                    },
                    "floor": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Floor"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "default": []
                    },
                    "images": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Images",
                        "default": []
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "created_at",
                    "updated_at"
                ],
                "title": "HouseResponse",
                "description": "\u623f\u6e90\u54cd\u5e94Schema\n\u7528\u4e8e\uff1a\u8fd4\u56de\u623f\u6e90\u8be6\u7ec6\u4fe1\u606f\uff08\u5305\u542b\u56fe\u7247\u8def\u5f84\uff09"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}